requirement.txt
amqp==5.1.0
asgiref==3.5.0
billiard==3.6.4.0
celery==5.2.6
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2
django-filter==21.1
djangorestframework==3.13.1
kombu==5.2.4
prompt-toolkit==3.0.29
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
vine==5.0.0
wcwidth==0.2.5
